TRELLO-W0VXgdcb: Change Event class back to interface
Rest-utils library and Verify Hub are passing in EventSinkHubEvent
object to Event Emitter via Event interface. This commit is needed to
avoid making code changes to existing code in Verify Hub and Rest Utils
library apart from upgrading its dependency to use latest version of
event emitter library.
