+++ src/test/java/uk/gov/ida/eventemitter/utils/EventMessageBuilder.java
@@ -2,12 +2,13 @@ package uk.gov.ida.eventemitter.utils;
 
 import uk.gov.ida.eventemitter.Event;
 import uk.gov.ida.eventemitter.EventDetailsKey;
+import uk.gov.ida.eventemitter.EventMessage;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 
-public class EventBuilder {
+public class EventMessageBuilder {
 
     private String eventType = "error";
     private String originatingService = "originating service";
@@ -19,17 +20,17 @@ public class EventBuilder {
         details.put(EventDetailsKey.error_id, UUID.randomUUID().toString());
     }
 
-    public static EventBuilder anEventMessage() {
-        return new EventBuilder();
+    public static EventMessageBuilder anEventMessage() {
+        return new EventMessageBuilder();
     }
 
-    public EventBuilder withDetailsField(final EventDetailsKey key, final String value) {
+    public EventMessageBuilder withDetailsField(final EventDetailsKey key, final String value) {
         details.put(key, value);
         return this;
     }
 
     public Event build() {
-        return new Event(
+        return new EventMessage(
             originatingService,
             sessionId,
             eventType,
